refactor get layout
