+++ packages/growth/src/inject.ts
@@ -1,6 +1,6 @@
-import { types as t } from '@onlook/parser';
-import { generate, parse, traverse } from '@onlook/parser';
+import { generate, parse, types as t, traverse } from '@onlook/parser';
 import { type FileOperations } from '@onlook/utility';
+import { getLayoutPath } from './helpers';
 import { builtWithScript } from './script';
 
 /**
@@ -14,25 +14,10 @@ export async function injectBuiltWithScript(
 ): Promise<boolean> {
     try {
         // Find the layout file - check both app/ and src/app/ directories
-        const possibleLayoutPaths = [
-            `${projectPath}/src/app/layout.tsx`,
-            `${projectPath}/app/layout.tsx`,
-        ];
-
-        let layoutPath: string | null = null;
-        for (const path of possibleLayoutPaths) {
-            const exists = await fileOps.fileExists(path);
-            if (exists) {
-                layoutPath = path;
-                break;
-            }
-        }
+        const layoutPath = await getLayoutPath(projectPath, fileOps.fileExists);
 
         if (!layoutPath) {
-            console.error(
-                'Layout file not found at any of the expected paths:',
-                possibleLayoutPaths,
-            );
+            console.error('Layout file not found');
             return false;
         }
